[Bronze V] Title: 이상한 기호, Time: 176 ms, Memory: 17720 KB, Score: 100 point -BaekjoonHub

+++ 백준/Bronze/15964. 이상한 기호/이상한 기호.java
@@ -1,4 +1,3 @@
-
 import java.util.Scanner;
 
 public class Main {
@@ -6,10 +5,10 @@ public class Main {
 
         Scanner sc = new Scanner(System.in);
         
-        int A = sc.nextInt();
-        int B = sc.nextInt();
+        long A = sc.nextLong();
+        long B = sc.nextLong();
         
-        int ans = (A+B) * (A-B);
+        long ans = (A+B) * (A-B);
         System.out.println(ans);
     }
 }
